Fix compilation warnings
git-svn-id: http://localhost:8080/svn/AlgoQuest/trunk@57 de6f410c-f7d5-3442-bb8a-973c24c52ac3

+++ AQUtil/aq/Utilities.cpp
@@ -358,7 +358,7 @@ void doubleToString( char* strVal, double dVal )
 }
 
 //------------------------------------------------------------------------------
-std::string getThesaurusFileName( char* path, int tableIdx, int columnIdx, int partIdx )
+std::string getThesaurusFileName( char* path, size_t tableIdx, size_t columnIdx, size_t partIdx )
 {
 	char szFN[ _MAX_PATH ];
 	if( path )
